Split cost calculator modules
